requirements.txt
# Core
fastapi
uvicorn

# LLMs & RAG
langchain
langchain-community
faiss-cpu
openai
ollama

# File Handling
pytesseract
PyPDF2
python-pptx

# Audio Processing
openai-whisper
pydub
# Utils
requests

# Dashboard & Analytics
plotly

# Async / Performance
aiofiles


uvicorn[standard]
pydantic
python-dotenv
python-docx
pandas
PyPDF2
Pillow
langdetect
